Make `ocpapimsubscriptionkey` optional; update job names
The `check_health_endpoint_fm_live.ps1` script was modified to make the `ocpapimsubscriptionkey` parameter optional by setting its default value to `$null`. A conditional check was added to ensure that the `Ocp-Apim-Subscription-Key` header is only added to the HTTP request if the `ocpapimsubscriptionkey` is not null.

In the `app-deploy.yml` file, the display names for the jobs `CheckFMNonLive` and `CheckFMLive` were updated to include the environment parameter, changing from "Check Fleet Manager Health" to "Check Fleet Manager Health - ${{parameters.Environment}}".