Make $ocpapimsubscriptionkey a mandatory parameter
The parameter `$ocpapimsubscriptionkey` was previously optional with a default value of `$null`. It has been changed to a mandatory parameter.

The code that conditionally added the `Ocp-Apim-Subscription-Key` header to the HTTP request only if `$ocpapimsubscriptionkey` was not null has been removed. Now, the header is always added to the HTTP request, assuming that `$ocpapimsubscriptionkey` is always provided due to its mandatory status.